packages/ml-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
python-dotenv==1.0.1
psycopg2-binary==2.9.10
redis==5.2.0
pandas==2.2.3
numpy==2.1.0
scikit-learn==1.5.2
xgboost==2.1.0
joblib==1.4.2
requests==2.32.3
aiohttp==3.11.0
sqlalchemy==2.0.36
pymongo==4.10.0
python-multipart==0.0.12
asyncpg==0.30.0
